backend/prediction.py

import cv2
import torch
import numpy as np
from models.model import YourModel  # Update with your actual model import

def predict_video(video_path):
    # Initialize model
    model = YourModel()
    model.eval()
    
    # Open video file
    cap = cv2.VideoCapture(video_path)
    predictions = []
    
    while cap.isOpened():
        ret, frame = cap.read()
        if not ret:
            break
            
        # Preprocess frame
        frame = cv2.resize(frame, (224, 224))
        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
        frame = torch.from_numpy(frame).float() / 255.0
        frame = frame.permute(2, 0, 1).unsqueeze(0)
        
        # Get prediction
        with torch.no_grad():
            output = model(frame)
            predictions.append(output.numpy().tolist())
    
    cap.release()
    
    # Process predictions to get final result
    final_prediction = process_predictions(predictions)
    return final_prediction

def process_predictions(predictions):
    # Implement your prediction processing logic here
    # This is a placeholder that you should replace with your actual logic
    return {
        "class": "your_predicted_class",
        "confidence": 0.95
    }
