backend/requirements.txt

flask==2.3.3
python-dotenv==1.0.0
supabase==1.0.3
opencv-python==4.8.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
